brain-fastapi/requirements.txt
fastapi
uvicorn[standard]
groq
langgraph
langchain
langchain-groq
langchain-core
pydantic
python-dotenv
feedparser
httpx
langfuse